docs: Clarify the user's path through the docs
Fixes #12

BREAKING CHANGE: This included moving all `cargo` stuff under a `cargo`
module in the public API.  The intent was to not add noise to the main
API with more tangential parts.  This might hide the extension trait too
much.  It might be added back to the root if this was found to confuse
people.