fix(cargo): Reuse the current target
I decided to go with this because I feel like correctness trumps
performance.

For those not specifying `--target`, the performance will surprise them.

For those specifying `--target`, the lack of correctness will surprise
them but is less easily detected.

Fixes #4, #44

BREAKING CHANGE: In this case, performance is a breaking change because
this might cause CI's to time out.