fix: tty test
When I test a program that uses `isatty`, it will detect that it is
not a tty even if it is a tty.
The cause was that the pipe was passed to stdin even when
`write_stdin` was not used.
Fixed to pass pipe to stdin only when `write_stdin`.